test template

+++ src/test2.hpp
@@ -82,7 +82,7 @@ namespace x2
     		}
     	private:
     		static int m_num_objects;
-    		static const int max_objects=3;
+    		static const size_t max_objects=3;
     		void init()
     		{
     			if( m_num_objects>max_objects)
